Consolidate background and navbar into layout component
