+++ components/Layout/index.ts
@@ -1,0 +1,1 @@
+export { Layout as default } from './Layout';
